Check if output is decorated
When the output is not decorated (`--no-ansi`), we can skip highlighting the error,
as it won't have any effect anyway.

The filenames will not be clickable and therefore we don't have to trim them.

Since `Output::isDecorated` was recently introduced in PHPStan we need to bump the minimum
requirement to `1.12.4`